changelog: Upcoming Features, Fraud, Mark when users with IAL2 of specific accounts are found
